Fix various UI bugs.
Three bugs fixed, all in `ui.c`:

**Bug 1 --- The actual crash:** `GEN_FrmGotoForm()` was setting `ui_current_form` but never setting `_pending_form_open = 1`. This meant `frmOpenEvent` was never delivered to `NewCommanderFormHandleEvent`. Without that event, `SetField()` was never called, so `NameH` (the handle used to read the commander name back) was never initialized. When the player eventually pressed a button, `GetField()` dereferenced that null handle → crash.

**Bug 2 --- Coordinate mapping:** The `_py2row()` function used `y * 14 / 80 / 8` which collapsed everything to row 0 for small y values. Replaced with `y * 28 / 160` which correctly maps the Palm's 160px height to 28 tile rows.

**Bug 3 --- Missing input handling:** `GEN_EvtGetEvent` only produced generic `keyDown` events, but `NewCommanderFormHandleEvent` only handles `ctlSelectEvent` (button presses). Added comprehensive form-specific joypad mapping for all major screens:

| Screen | A | B | C | Start | D-pad |
| --- | --- | --- | --- | --- | --- |
| New Commander | Confirm | --- | --- | --- | ↕ = cycle field, ←→ = change value |
| System Info | Warp | Buy Cargo | Sell Cargo | Shipyard | ↑=Special ↓=News |
| Encounter | Attack | Flee | Ignore | Interrupt | --- |
| Warp/Chart | Confirm | Back | --- | Shipyard | ↕ = scroll |
| Buy/Sell Cargo | Qty button | Back | --- | Shipyard | ↕ = scroll |

+++ Src/Encounter.c
@@ -63,6 +63,7 @@
 // *************************************************************************
 
 #include "external.h"
+#include "compat.h"
 
 static Boolean playerShipNeedsUpdate, opponentShipNeedsUpdate;
 
@@ -1382,7 +1383,7 @@ static void DrawEncounterForm()
 		DrawChars( SBuf, 6, 88 );
 	}			
 
-	d = sqrt( (int)(Ship.Tribbles/250) );
+	d = sqrt( Ship.Tribbles/250 );
 	for (i=0; i<d; ++i)
 	{
 		objindex = FrmGetObjectIndex( frmP, EncounterTribble0Button +
